Refactor(#30-battery-view): EditEnergyView 파일 분리
